Created the rest of controllers needed
Created the rest of controllers and verified the repositories

+++ src/main/java/com/marcels/pos/models/repositories/CategoryRepository.java
@@ -4,7 +4,7 @@ import com.marcels.pos.models.entities.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-@Repository 
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-
+@Repository
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
+    // Spring Data JPA ya te da el .findAll() y .save() automáticamente
 }
