+++ src/main/java/com/marcels/pos/models/repositories/PaymentMethodRepository.java
@@ -4,8 +4,6 @@ import com.marcels.pos.models.entities.PaymentMethod;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-@Repository 
-public interface PaymentMethodRepository extends JpaRepository<PaymentMethod, Long> {
+@Repository
+public interface PaymentMethodRepository extends JpaRepository<PaymentMethod, Integer> {
 }
-
-
